Restore hot reloading support Refactor TemplateFileProvider
